+++ app/code/community/Unirgy/DropshipPo/controllers/VendorController.php
@@ -239,7 +239,7 @@ class Unirgy_DropshipPo_VendorController extends Unirgy_Dropship_VendorControlle
                     ->setCarrierCode($_carrier)
                     ->setTitle($title);
 
-                $shipment->addTrack($track);
+                 $shipment->addTrack($track);
 
                 Mage::helper('udropship')->processTrackStatus($track, true, $isShipped);
 
@@ -325,32 +325,31 @@ class Unirgy_DropshipPo_VendorController extends Unirgy_Dropship_VendorControlle
 
             $session->setHighlight($highlight);
         } catch (Exception $e) {
-            $r = $this->getRequest();
-            $id = $r->getParam('id');
-            $po = Mage::getModel('udpo/po')->load($id);
-            Mage::helper('udropship')->assignVendorSkus($po);
-            Mage::helper('udropship/item')->hideVendorIdOption($po);
-            $store = Mage::app()->getStore();
+            $configValue = Mage::getStoreConfig('aftership_options/messages/aftership_validation');
+            //save, 422: repeated
+            if($configValue) {
+                $r = $this->getRequest();
+                $id = $r->getParam('id');
+                $po = Mage::getModel('udpo/po')->load($id);
+                Mage::helper('udropship')->assignVendorSkus($po);
+                Mage::helper('udropship/item')->hideVendorIdOption($po);
 
-            Mage::app()->setCurrentStore(Mage_Core_Model_App::ADMIN_STORE_ID);
-            foreach($po->getOrder() as $item){
-                $item->setQtyShipped(0);
-                //$item->setQtyToShip(1);
-                $item->setQtyCanceled(0);
-                $item->save();
+                $store = Mage::app()->getStore();
+                Mage::app()->setCurrentStore(Mage_Core_Model_App::ADMIN_STORE_ID);
+                foreach ($po->getOrder()->getItemsCollection()->getItems() as $item) {
+                    $item->setQtyShipped(0);
+                    $item->setQtyCanceled(0);
+                    $item->save();
+                }
+                foreach ($po->getItemsCollection()->getItems() as $item) {
+                    $item->setQtyShipped(0);
+                    $item->setQtyCanceled(0);
+                    $item->save();
+                }
+                $shipment->delete();
+                $po->save();
+                Mage::app()->setCurrentStore($store);
             }
-            $shipment->delete();
-            $po->save();
-//            $item->getOrder();
-//            foreach($udpo->getShipmentsCollection() as $item){
-//                if(!(count($item->getTracksCollection()->getItems())>1)){
-//                    $order = $item->getOrder()->getItems();
-//                    $item->delete();
-//                }
-//            }
-            $po->save();
-
-            Mage::app()->setCurrentStore($store);
             $session->addError($e->getMessage());
         }
